+++ src/shared/message/reply.rs
@@ -110,7 +110,7 @@ impl MessageReplyCommand {
         client: &mut EmailClient,
     ) -> Result<()> {
         let mailbox = self.mailbox.resolve(account)?;
-        let source = client.get_message(&mailbox, &self.id)?;
+        let source = client.get_message(&mailbox, &self.id, false)?;
 
         let raw = builder::build(
             BuilderArgs {